# Node 12 + upgrade hiredis