Hi!

We want to upgrade Node 12 LTS, and currently, hiredis-node doesn't support it.
This PR:
* Updates travis-ci configuration: to test again two latest LTS and stable, and remove old nodes configurations.
* Upgrade hiredis to 0.14.0 
* Support Node 12 due to deprecations

Notes:
* I ran this branch on our ci which includes lots of unit tests (there interacts with redis+hiredis) and full e2e tests and it passes.
* The "leaks" test is failing on my local computer (macOS Catalina 10.15), but I commented all tests and it's still failing so don't think it's related to my changes. 

Benchmarks:
```
# master

PING: 55248.62 ops/sec
SET: 59347.18 ops/sec
GET: 70671.38 ops/sec
LPUSH 8 bytes: 66225.17 ops/sec
LPUSH 64 bytes: 61162.08 ops/sec
LPUSH 512 bytes: 55096.42 ops/sec
LRANGE 10 elements, 8 bytes: 54794.52 ops/sec
LRANGE 100 elements, 8 bytes: 30911.90 ops/sec
LRANGE 100 elements, 64 bytes: 28490.03 ops/sec
LRANGE 100 elements, 512 bytes: 14847.81 ops/sec


# hiredis + node 12

PING: 61162.08 ops/sec
SET: 64308.68 ops/sec
GET: 74626.87 ops/sec
LPUSH 8 bytes: 69444.44 ops/sec
LPUSH 64 bytes: 71684.59 ops/sec
LPUSH 512 bytes: 70671.38 ops/sec
LRANGE 10 elements, 8 bytes: 58997.05 ops/sec
LRANGE 100 elements, 8 bytes: 30721.97 ops/sec
LRANGE 100 elements, 64 bytes: 28288.54 ops/sec
LRANGE 100 elements, 512 bytes: 16233.77 ops/sec
```