@yosiat 2019-11-21
Update: we deployed this code to production and it's working fine for us for the past two days.